template_tags: changing file names for testing and to ensure correct tag is called
