+++ adhocracy4/comments_async/templatetags/react_comments_async.py
@@ -1,7 +1,9 @@
 import json
 
 from django import template
+from django.conf import settings
 from django.contrib.contenttypes.models import ContentType
+from django.core.exceptions import ImproperlyConfigured
 from django.urls import reverse
 from django.utils.html import format_html
 
@@ -12,9 +14,8 @@ register = template.Library()
 
 
 @register.simple_tag(takes_context=True)
-def react_comments_async(context, obj):
+def react_comments_async(context, obj, debate_app=True):
     request = context['request']
-
     user = request.user
     is_authenticated = bool(user.is_authenticated)
     is_moderator = user.is_superuser or user in obj.project.moderators.all()
@@ -27,16 +28,24 @@ def react_comments_async(context, obj):
     has_comment_permission = user.has_perm(permission, obj)
 
     would_have_comment_permission = NormalUser().would_have_perm(
-        permission, obj
-    )
+        permission, obj)
+
+    comments_contenttype = ContentType.objects.get_for_model(Comment)
+    pk = obj.pk
+
+    module_type = bool(debate_app)
 
     comments_api_url = reverse('comments-list',
                                kwargs={'content_type': contenttype.pk,
                                        'object_pk': obj.pk}
                                )
 
-    comments_contenttype = ContentType.objects.get_for_model(Comment)
-    pk = obj.pk
+    comment_category_choices = getattr(settings, 'A4_COMMENT_CATEGORIES', None)
+    if comment_category_choices:
+        comment_category_choices = dict(
+            (x, str(y)) for x, y in comment_category_choices)
+    else:
+        raise ImproperlyConfigured('set A4_COMMENT_CATEGORIES in settings')
 
     attributes = {
         'commentsApiUrl': comments_api_url,
@@ -46,13 +55,14 @@ def react_comments_async(context, obj):
         'isAuthenticated': is_authenticated,
         'isModerator': is_moderator,
         'user_name': user_name,
-        'isReadOnly': (not has_comment_permission and
-                       not would_have_comment_permission),
-        'anchoredCommentId': anchoredCommentId
+        'isReadOnly': (not has_comment_permission
+                       and not would_have_comment_permission),
+        'commentCategoryChoices': comment_category_choices,
+        'anchoredCommentId': anchoredCommentId,
+        'moduleType': module_type
     }
 
     return format_html(
-        '<div data-a4-widget="comment_async" '
+        '<div data-a4-widget="comment_async_with_categories" '
         'data-attributes="{attributes}"></div>',
-        attributes=json.dumps(attributes)
-    )
+        attributes=json.dumps(attributes))
